refactor: integrate FilmMapper into FilmService
- Updated `create`, `update`, and `findById` methods to handle DTO conversion.
- Refactored `findAll` to return a collection of `FilmDto` using Java Streams.
- Encapsulated model-to-dto transformation logic within `FilmMapper`.
- Note: Like-related functionality remains unchanged for separate DB migration.

+++ src/main/java/ru/yandex/practicum/filmorate/dto/film/FilmCreateRequest.java
@@ -27,7 +27,7 @@ public class FilmCreateRequest {
     @NotNull(message = "Возрастной рейтинг должен быть указан")
     @Min(message = "Неверно указан id возрастного рейтинга", value = 1L)
     @Max(message = "Неверно указан id возрастного рейтинга", value = 5L)
-    private Long ratingId;
+    private Integer ratingId;
 
     @NotEmpty(message = "Жанры фильма должны быть указаны")
     List<@NotNull @Positive Integer> genreIds;
